Completed SpaceShip Exam
Completed with the detailed log of the days and the amount of resources mined per day.

Fun stuff: the names of the planets have been taken out  from "https://www.fantasynamegenerators.com/planet_names.php" . THis is done as a generator class of names. THey shoould not be repeated as long as the amount of planets to be visited is not the amount of planets available. The safe methods with % should be checked out to reference ways of making stuff that won't get a NullPointerException